给予了默认值显示，修复了图标显示错误bug

+++ Assets/Shili/Codes/CanvasCode/PropsPoolUICanvas.cs
@@ -22,7 +22,7 @@ public class PropsPoolUICanvas : MonoBehaviour
             if (go1.TryGetComponent<Image>(out var img1) && PropManager.Instance.TryGetPropConfig<BasePropConfig>(allEPropTypes[i+1], out var propConfig1))
             {
                 img1.sprite = propConfig1.PropSprite;
-                go1.GetComponent<PropChoiceButtonSet>().id = (int)allEPropTypes[i+1];
+                go1.GetComponent<PropChoiceButtonSet>().id = allEPropTypes[i+1];
             }
         }
         GameObject go = Instantiate(prefab);
@@ -32,7 +32,7 @@ public class PropsPoolUICanvas : MonoBehaviour
         if (go.TryGetComponent<Image>(out var img) && PropManager.Instance.TryGetPropConfig<BasePropConfig>(allEPropTypes[0], out var propConfig))
         {
             img.sprite = propConfig.PropSprite;
-            go.GetComponent<PropChoiceButtonSet>().id =(int) allEPropTypes[0];
+            go.GetComponent<PropChoiceButtonSet>().id =allEPropTypes[0];
         }
     }
     public void SetOnDisable()
